requirements.txt
langchain==0.0.300
openai==0.28.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0